+++ src/Sulu/Component/Security/Authorization/AccessControl/AccessControlManagerInterface.php
@@ -1,0 +1,34 @@
+<?php
+/*
+ * This file is part of the Sulu CMF.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\Security\Authorization\AccessControl;
+
+/**
+ * Defines methods for assigning security to objects
+ */
+interface AccessControlManagerInterface
+{
+    /**
+     * Sets the permissions for the object with the given class and id for the given security identity
+     * @param string $type The name of the class to protect
+     * @param string $identifier
+     * @param string $securityIdentity
+     * @param $permissions
+     */
+    public function setPermissions($type, $identifier, $securityIdentity, $permissions);
+
+    /**
+     * Returns the permissions for all security identities
+     * @param string $type The type of the protected object
+     * @param string $identifier The identifier of the protected object
+     * @return array
+     */
+    public function getPermissions($type, $identifier);
+}